feat: 경매는 진행 중 상태가 아니면 사용자의 요청을 처리할 수 없다. (#85)
* fix: AuctionInfo의 재고수량 타입을 변경한다. int -> long

* feat: 경매 입찰(구매)시 진행 중인 경매인지 검증

- 진행 중인 경매가 아닌 경우 A016 에러를 반환한다.
- 경매 상태 비교를 Service에서 하지 않고, 경매상태(AuctionStatus)에게 직접 물어보도록 수정한다.
- 경매 상태에 의존적인 '입찰 성공' 테스트를 일시적으로 Disabled 한다.

* feat: Auction의 updateStatus를 호출하여 상태 업데이트

- TODO 임시로직: 추후 스케줄러 등 믿을 수 있는 경매 관리자가 상태를 업데이트 하도록 수정 필요
- 현재는 현재 시간 기준과 재고를 기준으로 판단하는 로직을 활용한다.

* feat: 경매 입찰(구매)시 진행 중인 경매인지 검증

- 진행 중인 경매가 아닌 경우 A016 에러를 반환한다.
- 경매 상태 비교를 Service에서 하지 않고, 경매상태(AuctionStatus)에게 직접 물어보도록 수정한다.

* feat: 경매는 자기 자신을 고유하게 구분하는 id를 가진다.

* feat: 경매 재고 및 인당 구매 수량 제한 타입 변경 int -> long

* feat: 경매(Auction) Builder에서 경매상태를 받을 수 있다.

- 전달되는 AuctionStatus가 Null인 경우 기본값으로 AuctionStatus.WAITING을 사용한다.

* test: 테스트를 위한 FakeAuctionRepository를 만들고 활용

- AuctionRepository 인터페이스를 구현하여 활용한다.

* fix: 잘못들어가있던 println 구문 제거

* feat: Auction 생성자에서 경매 상태를 지정할 수 없도록 수정

- AuctionStatus는 WAITING을 기본값으로 가진다.
- Auction의 ID는 setter를 통해 지정할 수 있도록 수정한다.

---------

+++ src/main/java/com/wootecam/luckyvickyauction/core/auction/dto/AuctionInfo.java
@@ -22,7 +22,7 @@ public record AuctionInfo(
         long originPrice,
         long currentPrice,
         long stock,
-        int maximumPurchaseLimitCount,
+        long maximumPurchaseLimitCount,
         boolean isShowStock,
         AuctionStatus status
 ) {
@@ -71,7 +71,7 @@ public record AuctionInfo(
         }
     }
 
-    private void validateMaximumPurchaseLimitCount(int maximumPurchaseLimitCount) {
+    private void validateMaximumPurchaseLimitCount(long maximumPurchaseLimitCount) {
         if (maximumPurchaseLimitCount <= 0) {
             throw new BadRequestException(String.format(ERROR_MAXIMUM_PURCHASE_LIMIT_COUNT, maximumPurchaseLimitCount),
                     ErrorCode.A003);
